+++ internal/config/config.go
@@ -162,6 +162,8 @@ type AnalyzeConfig struct {
 	DetectErrorSuppressionPropagation        bool     `toml:"detect_error_suppression_propagation"`
 	DetectLoopInvariantExcelObjectResolution bool     `toml:"detect_loop_invariant_excel_object_resolution"`
 	DetectUnsafeSQLConstruction              bool     `toml:"detect_unsafe_sql_construction"`
+	DetectRiskyModuleState                   bool     `toml:"detect_risky_module_state"`
+	DetectRedimPreserveInLoops               bool     `toml:"detect_redim_preserve_in_loops"`
 }
 
 type lintRuleAdapter struct {
@@ -248,6 +250,8 @@ var analyzeRuleAdapters = map[string]analyzeRuleAdapter{
 	"VBA237": {Get: func(c AnalyzeConfig) bool { return c.DetectErrorSuppressionPropagation }, Set: func(c *AnalyzeConfig, v bool) { c.DetectErrorSuppressionPropagation = v }},
 	"VBA238": {Get: func(c AnalyzeConfig) bool { return c.DetectLoopInvariantExcelObjectResolution }, Set: func(c *AnalyzeConfig, v bool) { c.DetectLoopInvariantExcelObjectResolution = v }},
 	"VBA239": {Get: func(c AnalyzeConfig) bool { return c.DetectUnsafeSQLConstruction }, Set: func(c *AnalyzeConfig, v bool) { c.DetectUnsafeSQLConstruction = v }},
+	"VBA240": {Get: func(c AnalyzeConfig) bool { return c.DetectRiskyModuleState }, Set: func(c *AnalyzeConfig, v bool) { c.DetectRiskyModuleState = v }},
+	"VBA241": {Get: func(c AnalyzeConfig) bool { return c.DetectRedimPreserveInLoops }, Set: func(c *AnalyzeConfig, v bool) { c.DetectRedimPreserveInLoops = v }},
 }
 
 var (